Use Docker API to pull images
Replace a shell command with a call to the Docker API to pull images for
the Docker executor.

Also remove unused code as well as rename `SetupAllSuite` functions to
the testify `SetupSuite` function, so that testify will call them when
the suite is initially created.

Fixes #283
