+++ pkg/test/devstack/concurrency_test.go
@@ -31,7 +31,7 @@ func TestDevstackConcurrencySuite(t *testing.T) {
 }
 
 // Before all suite
-func (suite *DevstackConcurrencySuite) SetupAllSuite() {
+func (suite *DevstackConcurrencySuite) SetupSuite() {
 
 }
 
@@ -45,7 +45,7 @@ func (suite *DevstackConcurrencySuite) TearDownTest() {
 
 }
 
-func (suite *DevstackConcurrencySuite) TearDownAllSuite() {
+func (suite *DevstackConcurrencySuite) TearDownSuite() {
 
 }
