+++ pkg/test/devstack/publish_on_error_test.go
@@ -31,7 +31,7 @@ func TestPublishOnErrorSuite(t *testing.T) {
 }
 
 // Before all suite
-func (s *PublishOnErrorSuite) SetupAllSuite() {
+func (s *PublishOnErrorSuite) SetupSuite() {
 
 }
 
@@ -44,7 +44,7 @@ func (s *PublishOnErrorSuite) SetupTest() {
 func (suite *PublishOnErrorSuite) TearDownTest() {
 }
 
-func (s *PublishOnErrorSuite) TearDownAllSuite() {
+func (s *PublishOnErrorSuite) TearDownSuite() {
 
 }
