+++ pkg/test/devstack/sharding_test.go
@@ -36,7 +36,7 @@ func TestShardingSuite(t *testing.T) {
 }
 
 // Before all suite
-func (suite *ShardingSuite) SetupAllSuite() {
+func (suite *ShardingSuite) SetupSuite() {
 
 }
 
@@ -49,7 +49,7 @@ func (suite *ShardingSuite) SetupTest() {
 func (suite *ShardingSuite) TearDownTest() {
 }
 
-func (suite *ShardingSuite) TearDownAllSuite() {
+func (suite *ShardingSuite) TearDownSuite() {
 
 }
