+++ pkg/test/devstack/util_ulimit.go
@@ -12,7 +12,8 @@ import (
 
 const minimumOpenFiles int = 512
 
-func shouldRunShardingTest() (bool, error) { //nolint:unused
+//nolint:unused // golangci-lint complains that this is unused, but it's in pkg/test/devstack/sharding_test.go
+func shouldRunShardingTest() (bool, error) {
 	ulimitValue := 0
 
 	if _, err := exec.LookPath("ulimit"); err == nil {
