+++ pkg/test/devstack/util_ulimit_windows.go
@@ -1,5 +1,6 @@
 package devstack
 
-func shouldRunShardingTest() (bool, error) { //nolint:unused
+//nolint:unused // golangci-lint complains that this is unused, but it's in pkg/test/devstack/sharding_test.go
+func shouldRunShardingTest() (bool, error) {
 	return true, nil
 }
